pyproject.toml
[tool.poetry]
name = "ghostpost_backend"
version = "0.1.0"
description = ""
authors = ["sohail aslam <sohailaslam0707@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
djangorestframework = "^3.12.0"
django-cors-headers = "^3.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
